strands-agents-examples/90-deployment/20-fast-api/requirements.txt
strands-agents==1.13.0
strands-agents-tools==0.2.12

pydantic==2.12.3
fastapi[standard]==0.119.1
